# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almaz",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/puzzle-codes/verify": {
            "post": {
                "tags": [
                    "POST"
                ],
                "summary": "Проверить код",
                "operationId": "verify",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/puzzle_code_verify"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Элемент создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/puzzle_code_verify_created"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неправильный параметр"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "puzzle_code_verify": {
                "example": {
                    "code": "VDR-XZT-IAN"
                }
            },
            "puzzle_code_verify_created": {
                "example": {
                    "status": "success",
                    "response": {
                        "code": [
                            "Введен корректный код"
                        ]
                    },
                    "color": "popArt",
                    "size": "20x30"
                }
            },
...
        }
    }
}